+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Version 1.3.0 (unreleased) 
   - support CIDR ranges in `server.rules.admin-ips`, in addition to individual IP addresses
+  - log invalid `server.rules.admin-ips` entries at warn level, even if the rules module is configured with a higher log level
   - fix account manager completed metric OnGenerate error
   - import latest go-eth2-wallet-store-s3 to enable force-path-style on S3